add interface and combine in one compoenet
